requirements.txt
haystack-ai==2.7.0
python-dotenv==1.0.1
fastapi==0.115.6
pgvector-haystack==1.2.0
sentence-transformers==3.3.1